server/pyproject.toml
[tool.poetry]
name = "iot-server"
version = "0.1.0"
description = "iot-device server"
authors = []

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
flask = "3.0.2"
pyserial = "3.5"
toml = "0.10.2"

[tool.poetry.dev-dependencies]
pytest = "8.0.2"
black = "22.3.0"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
